pickle-factory/requirements.txt
cassandra-driver==3.11.0
lz4==0.10.1
scales==1.0.9
